progress of making things work

+++ Salvavida/Package/Runtime/ObservableArray.cs
@@ -1,6 +1,5 @@
 using Salvavida.DefaultImpl;
 using System;
-using System.Buffers;
 using System.Collections;
 using System.Collections.Generic;
 
@@ -13,20 +12,23 @@ namespace Salvavida
 
         }
 
-        public ObservableArray(string propName, T?[] src, bool saveSeparately)
+        public ObservableArray(string propName, T?[]? src, bool saveSeparately)
             : base(propName, saveSeparately)
         {
-            _arr = default!;
             SwapSource(src, false);
         }
 
-        private T?[] _arr;
+        private T?[]? _arr;
+
+        public override bool IsDirty => _arr != null && base.IsDirty;
 
         public T? this[int index]
         {
-            get => _arr[index];
+            get => _arr == null ? default : _arr[index];
             set
             {
+                if (_arr == null)
+                    throw new NullReferenceException(nameof(_arr));
                 var oldVal = _arr[index];
                 _arr[index] = value;
                 OnItemSet(value, index);
@@ -35,21 +37,21 @@ namespace Salvavida
             }
         }
 
-        object? IList.this[int index] { get => _arr[index]; set => this[index] = (T?)value; }
+        object? IList.this[int index] { get => this[index]; set => this[index] = (T?)value; }
 
-        public int Count => _arr.Length;
+        public int Count => _arr?.Length ?? 0;
 
         public bool IsReadOnly => false;
 
         bool IList.IsFixedSize => true;
 
-        bool ICollection.IsSynchronized => _arr.IsSynchronized;
+        bool ICollection.IsSynchronized => _arr?.IsSynchronized ?? false;
 
-        object ICollection.SyncRoot => _arr.SyncRoot;
+        object? ICollection.SyncRoot => _arr?.SyncRoot ?? null;
 
-        public T?[] RetrieveSource() => _arr;
+        public T?[]? RetrieveSource() => _arr;
 
-        public object RetrieveSourceRaw() => _arr;
+        public object? RetrieveSourceRaw() => _arr;
 
         public Type CollectionType => typeof(T?[]);
 
@@ -57,7 +59,7 @@ namespace Salvavida
         {
             if (!SaveSeparately)
                 return;
-            serializer.SaveArray(_arr, ctx, SvId);
+            serializer.SaveObject(_arr, ctx, SvId, PathBuilder.Type.Property);
         }
 
         public override void Deserialize(Serializer serializer, SerializeContext ctx)
@@ -68,13 +70,13 @@ namespace Salvavida
             SwapSource(arr);
         }
 
-        public void SwapSource(T?[] array)
+        public void SwapSource(T?[]? array)
         {
             _isDirty = true;
             SwapSource(array, true);
         }
 
-        private void SwapSource(T?[] array, bool notifyChanges)
+        private void SwapSource(T?[]? array, bool notifyChanges)
         {
             if (_arr != null)
             {
@@ -115,6 +117,8 @@ namespace Salvavida
 
         protected override CollectionChangeInfo<T?> CreateSaveAllEvent()
         {
+            if (_arr == null)
+                throw new NullReferenceException(nameof(_arr));
             return CollectionChangeInfo<T?>.Add(_arr, 0);
         }
 
@@ -133,15 +137,15 @@ namespace Salvavida
         public bool Contains(T? item) => Array.IndexOf(_arr, item) >= 0;
         bool IList.Contains(object value) => Contains((T?)value);
 
-        public ArrayEnumerator GetEnumerator() => new(_arr);
+        public ArrayEnumerator GetEnumerator() => _arr == null ? throw new NullReferenceException(nameof(_arr)) : new(_arr);
 
         IEnumerator<T?> IEnumerable<T?>.GetEnumerator() => GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
-        void ICollection<T?>.CopyTo(T?[] array, int arrayIndex) => _arr.CopyTo(array, arrayIndex);
+        void ICollection<T?>.CopyTo(T?[] array, int arrayIndex) => _arr?.CopyTo(array, arrayIndex);
 
-        void ICollection.CopyTo(Array array, int index) => _arr.CopyTo(array, index);
+        void ICollection.CopyTo(Array array, int index) => _arr?.CopyTo(array, index);
 
         public int IndexOf(T? item) => Array.IndexOf(_arr, item);
 
@@ -212,14 +216,13 @@ namespace Salvavida
     public sealed class ObservableArraySavable<T> : ObservableCollectionSavable<ObservableArraySavable<T>, T>, IList<T?>, IReadOnlyList<T?>, IList, IEnumerable<T?>, IEnumerable, ICollectionWrapper<T?[]>
         where T : ISavable
     {
-        public ObservableArraySavable(string propName, T?[] src, bool saveSeparately)
+        public ObservableArraySavable(string propName, T?[]? src, bool saveSeparately)
             : base(propName, saveSeparately)
         {
-            _arr = default!;
             SwapSource(src, false);
         }
 
-        private T?[] _arr;
+        private T?[]? _arr;
         private string[]? _idsOnDeserialized;
 
         public override bool IsDirty
@@ -244,9 +247,11 @@ namespace Salvavida
 
         public T? this[int index]
         {
-            get => _arr[index];
+            get => _arr == null ? throw new NullReferenceException(nameof(_arr)) : _arr[index];
             set
             {
+                if (_arr == null)
+                    throw new NullReferenceException(nameof(_arr));
                 var oldVal = _arr[index];
                 _arr[index] = value;
                 OnItemSet(value, index);
@@ -255,21 +260,21 @@ namespace Salvavida
             }
         }
 
-        object? IList.this[int index] { get => _arr[index]; set => this[index] = (T?)value; }
+        object? IList.this[int index] { get => this[index]; set => this[index] = (T?)value; }
 
-        public int Count => _arr.Length;
+        public int Count => _arr?.Length ?? 0;
 
         public bool IsReadOnly => false;
 
         bool IList.IsFixedSize => true;
 
-        bool ICollection.IsSynchronized => _arr.IsSynchronized;
+        bool ICollection.IsSynchronized => _arr?.IsSynchronized ?? false;
 
-        object ICollection.SyncRoot => _arr.SyncRoot;
+        object? ICollection.SyncRoot => _arr?.SyncRoot ?? null;
 
-        public T?[] RetrieveSource() => _arr;
+        public T?[]? RetrieveSource() => _arr;
 
-        public object RetrieveSourceRaw() => _arr;
+        public object? RetrieveSourceRaw() => _arr;
 
         public Type CollectionType => typeof(T[]);
 
@@ -280,19 +285,33 @@ namespace Salvavida
             var tempIds = SvHelper.idListPool.Get();
             try
             {
-                for (var i = 0; i < _arr.Length; i++)
+                if (_arr != null)
                 {
-                    var elem = _arr[i];
-                    serializer.Save(elem, ctx, PathBuilder.Type.Collection);
-                    tempIds.Add(elem.SvId);
-                }
-                serializer.SaveList(tempIds, ctx, SvHelper.PROPNAME_COLLECTION_METADATA);
-                for (var i = 0; i < _idsOnDeserialized.Length; i++)
-                {
-                    var oldId = _idsOnDeserialized[i];
-                    if (tempIds.IndexOf(oldId) < 0)
+                    for (var i = 0; i < _arr.Length; i++)
                     {
-                        serializer.DeleteObject(ctx, oldId, PathBuilder.Type.Collection);
+                        var elem = _arr[i];
+                        if (elem == null)
+                            continue;
+                        if (elem.SvId == null)
+                            throw new ArgumentNullException("elem.SvId");
+                        serializer.Save(elem, ctx, PathBuilder.Type.Collection);
+                        tempIds.Add(elem.SvId);
+                    }
+                    serializer.SaveObject(tempIds, ctx, SvHelper.PROPNAME_COLLECTION_METADATA, PathBuilder.Type.Property);
+                }
+                else
+                {
+                    serializer.DeleteObject(ctx, SvHelper.PROPNAME_COLLECTION_METADATA, PathBuilder.Type.Property);
+                }
+                if (_idsOnDeserialized != null)
+                {
+                    for (var i = 0; i < _idsOnDeserialized.Length; i++)
+                    {
+                        var oldId = _idsOnDeserialized[i];
+                        if (tempIds.IndexOf(oldId) < 0)
+                        {
+                            serializer.DeleteObject(ctx, oldId, PathBuilder.Type.Collection);
+                        }
                     }
                 }
             }
@@ -308,8 +327,8 @@ namespace Salvavida
         {
             if (!SaveSeparately)
                 return;
-            var tempIds = serializer.ReadObject<string?[]?>(ctx, SvHelper.PROPNAME_COLLECTION_METADATA, PathBuilder.Type.Collection);
-            T[] arr;
+            var tempIds = serializer.ReadObject<string[]?>(ctx, SvHelper.PROPNAME_COLLECTION_METADATA, PathBuilder.Type.Collection);
+            T?[]? arr;
             if (tempIds == null || tempIds.Length == 0)
                 arr = null;
             else
@@ -325,13 +344,13 @@ namespace Salvavida
         }
 
 
-        public void SwapSource(T?[] array)
+        public void SwapSource(T?[]? array)
         {
             _isDirty = true;
             SwapSource(array, true);
         }
 
-        private void SwapSource(T?[] array, bool notifyChanges)
+        private void SwapSource(T?[]? array, bool notifyChanges)
         {
             if (_arr != null)
             {
@@ -370,6 +389,8 @@ namespace Salvavida
 
         protected override CollectionChangeInfo<T?> CreateSaveAllEvent()
         {
+            if (_arr == null)
+                throw new NullReferenceException(nameof(_arr));
             return CollectionChangeInfo<T?>.Add(_arr, 0);
         }
 
@@ -377,15 +398,15 @@ namespace Salvavida
 
         bool IList.Contains(object value) => Contains((T?)value);
 
-        public ArrayEnumerator GetEnumerator() => new(_arr);
+        public ArrayEnumerator GetEnumerator() => _arr == null ? throw new NullReferenceException(nameof(_arr)) : new(_arr);
 
         IEnumerator<T?> IEnumerable<T?>.GetEnumerator() => GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
-        void ICollection<T?>.CopyTo(T?[] array, int arrayIndex) => _arr.CopyTo(array, arrayIndex);
+        void ICollection<T?>.CopyTo(T?[] array, int arrayIndex) => _arr?.CopyTo(array, arrayIndex);
 
-        void ICollection.CopyTo(Array array, int index) => _arr.CopyTo(array, index);
+        void ICollection.CopyTo(Array array, int index) => _arr?.CopyTo(array, index);
 
         public int IndexOf(T? item) => Array.IndexOf(_arr, item);
